D3ASIM-968: Refactored bid-related common tasks and events outside of the strategies and the BaseStrategy to a new BidUpdatedStrategy class.
