+++ tests/test_strategy_base.py
@@ -21,7 +21,7 @@ import pendulum
 
 from d3a.constants import TIME_ZONE
 from d3a.d3a_core.exceptions import MarketException
-from d3a.models.strategy import BaseStrategy, Offers
+from d3a.models.strategy import BidEnabledStrategy, Offers
 from d3a.models.market.market_structures import Offer, Trade, Bid
 
 
@@ -155,7 +155,7 @@ def offer_to_accept():
 
 @pytest.fixture
 def base():
-    base = BaseStrategy()
+    base = BidEnabledStrategy()
     base.owner = FakeOwner()
     base.area = FakeArea()
     return base
